requirements.txt
asgiref==3.4.1
bleach==4.1.0
certifi==2021.10.8
charset-normalizer==2.0.7
dj-database-url==0.5.0
Django==3.2.7
django-crispy-forms==1.13.0
django-environ==0.8.0
django-heroku==0.3.1
django-markdownify==0.9.0
django-multiselectfield==0.1.12
django-on-heroku==1.1.2
django-postgres-extensions==0.9.3
django-utils-six==2.0
djangorestframework==3.12.4
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.8.2
Markdown==3.3.4
packaging==21.2
psycopg2==2.9.2
pyparsing==2.4.7
pytz==2021.3
requests==2.26.0
six==1.16.0
sqlparse==0.4.2
typing-extensions==3.10.0.2
urllib3==1.26.7
webencodings==0.5.1
whitenoise==5.3.0
zipp==3.6.0
